Claim function and test cases added + Channel State Cycle function fixed

+++ contracts/EPNSCore/EPNSCoreProxy.sol
@@ -5,8 +5,6 @@ pragma solidity ^0.6.0;
 import "@openzeppelin/contracts/proxy/TransparentUpgradeableProxy.sol";
 
 contract EPNSCoreProxy is TransparentUpgradeableProxy {
-
-
     constructor(
         address _logic,
         address _governance,
@@ -17,7 +15,24 @@ contract EPNSCoreProxy is TransparentUpgradeableProxy {
         address _lendingPoolProviderAddress,
         address _daiAddress,
         address _aDaiAddress,
-        uint _referralCode
-    ) public payable TransparentUpgradeableProxy(_logic, _governance, abi.encodeWithSignature('initialize(address,address,address,address,address,address,address,uint256)', _pushChannelAdmin, _pushTokenAddress, _wethAddress, _uniswapRouterAddress, _lendingPoolProviderAddress, _daiAddress, _aDaiAddress,_referralCode)) {}
-
+        uint256 _referralCode
+    )
+        public
+        payable
+        TransparentUpgradeableProxy(
+            _logic,
+            _governance,
+            abi.encodeWithSignature(
+                "initialize(address,address,address,address,address,address,address,uint256)",
+                _pushChannelAdmin,
+                _pushTokenAddress,
+                _wethAddress,
+                _uniswapRouterAddress,
+                _lendingPoolProviderAddress,
+                _daiAddress,
+                _aDaiAddress,
+                _referralCode
+            )
+        )
+    {}
 }
